% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv.R
\name{nv}
\alias{nv}
\title{convenience function: make a 'named vector'}
\usage{
nv(x)
}
\arguments{
\item{x}{a character vector, or something that can become one}
}
\value{
\preformatted{ the same vector, but with its own values as names 
}
}
\description{
This trivial function is mostly for convenience in iterating over vectors.
As a side effect, it exposes any duplicates in supposedly unique indices.
}
\examples{
  vec <- letters[1:5]
  vec
  nv(vec) 

}
