% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_check.R
\name{tbk_check}
\alias{tbk_check}
\title{check that a file is actually a tabixmate file}
\usage{
tbk_check(tbk_file)
}
\arguments{
\item{tbk_file}{character string, the file}
}
\value{
\preformatted{     logical, is it or isn't it 
}
}
\description{
check that a file is actually a tabixmate file
}
