% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbmater.R
\name{tbk_constants}
\alias{tbk_constants}
\alias{tbmate_constants}
\title{constants for tbmate:}
\description{
MAX_DOUBLE16   (32766)
HDR_ID         (3)
HDR_VERSION    (4)
HDR_DATA_TYPE  (8)
HDR_MAX_OFFSET (8)
HDR_EXTRA      (8169)
HDR_TOTALBYTES (8192)
}
