% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_data.R
\name{tbk_data}
\alias{tbk_data}
\title{Get data from tbk}
\usage{
tbk_data(
  tbk_fnames,
  idx_fname = NULL,
  probes = NULL,
  show.unaddressed = FALSE,
  chrm = NULL,
  beg = NULL,
  end = NULL,
  as.matrix = FALSE,
  simplify = FALSE,
  name.use.base = TRUE,
  max_addr = 3000,
  max_source = 10^6,
  max_pval = 0.05,
  min_coverage = 5,
  all_units = FALSE
)
}
\arguments{
\item{idx_fname}{index file. If not given, use the idx.gz in first path}

\item{probes}{probe names}

\item{simplify}{reduce matrix to vector if only one sample is queried}

\item{name.use.base}{use basename for sample name}

\item{max_addr}{random addressing if under max_addr}

\item{max_pval}{maximum sig2 for float.float}

\item{min_coverage}{minimum sig2 for float.int}

\item{all_units}{retrieve all units for float.float and float.int}

\item{fnames}{tbk_fnames}

\item{min_source}{random addressing if source size is under min_source}
}
\value{
\preformatted{         a numeric matrix
}
}
\description{
Assumptions:
\enumerate{
\item All the tbks have the same index.
\item If idx_fname not given, idx.gz is in the same folder as the first sample.
}
}
