% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_data0.R
\name{tbk_data0}
\alias{tbk_data0}
\title{initial tabixmate file read}
\usage{
tbk_data0(
  tbk_fnames,
  idx_addr,
  all_units = FALSE,
  max_addr = 3000,
  max_source = 10^6
)
}
\arguments{
\item{tbk_fnames}{filenames}

\item{idx_addr}{index addresses}

\item{all_units}{all units? (FALSE)}

\item{max_addr}{maximum address (3000)}

\item{max_source}{maximimum source size (1000000)}
}
\description{
initial tabixmate file read
}
