% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_data_addr.R
\name{tbk_data_addr}
\alias{tbk_data_addr}
\title{tabixmate data address}
\usage{
tbk_data_addr(tbk_fnames, idx_addr, all_units = FALSE)
}
\arguments{
\item{tbk_fnames}{filenames}

\item{idx_addr}{index addresses}

\item{all_units}{all units? (FALSE)}
}
\description{
tabixmate data address
}
