% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_data_bulk.R
\name{tbk_data_bulk}
\alias{tbk_data_bulk}
\title{bulk read from tabixmate files}
\usage{
tbk_data_bulk(tbk_fnames, idx_addr, all_units = FALSE)
}
\arguments{
\item{tbk_fnames}{tabixmate filenames}

\item{idx_addr}{tabixmate index addresses}

\item{all_units}{all units? (FALSE)}
}
\description{
bulk read from tabixmate files
}
