% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_hdr.R
\name{tbk_hdr}
\alias{tbk_hdr}
\title{read a tbk file header}
\usage{
tbk_hdr(tbk_file)
}
\arguments{
\item{tbk_file}{the filename}
}
\value{
a 'tbk' S3 object with fields tbk_version, dtype, smax, num, and msg
}
\description{
read a tbk file header
}
