% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_hdr_write.R
\name{tbk_hdr_write}
\alias{tbk_hdr_write}
\title{like it says: write a tabixmate header}
\usage{
tbk_hdr_write(out, data, dtype = "FLOAT", msg = "", tbk_version = 1)
}
\arguments{
\item{out}{filename}

\item{data}{the data}

\item{dtype}{data type ("FLOAT")}

\item{msg}{the message ("")}

\item{tbk_version}{tbmate file format version (1)}
}
\value{
the result of writeBin
}
\description{
like it says: write a tabixmate header
}
