% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_hdrs.R
\name{tbk_hdrs}
\alias{tbk_hdrs}
\title{retrieve tbk headers}
\usage{
tbk_hdrs(tbk_fnames)
}
\arguments{
\item{tbk_fnames}{the tbk file names}
}
\value{
an S3 object of class tbk
}
\description{
retrieve tbk headers
}
