% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_pack.R
\name{tbk_pack}
\alias{tbk_pack}
\title{Pack data to tbk}
\usage{
tbk_pack(
  data,
  data2 = NULL,
  out_dir = NULL,
  out_fname = NULL,
  dtype = "FLOAT",
  idx_fname = NULL,
  tbk_version = 1,
  msg = "",
  na.token = -1,
  link_idx = FALSE
)
}
\arguments{
\item{data}{the data}

\item{data2}{2nd data if float.float or float.int}

\item{out_dir}{output dir (out_fname is set to out_dir/colnames(data).tbk)}

\item{out_fname}{output tbk file name}

\item{dtype}{data type: INT1,INT2,INT32,FLOAT,DOUBLE,STRINGD, or STRINGF}

\item{idx_fname}{index file name}

\item{link_idx}{whether or not a link is generated to the index}
}
\description{
Pack data to tbk
}
