% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_read_chunk.R
\name{tbk_read_chunk}
\alias{tbk_read_chunk}
\title{read a chunk of a tabixmate file}
\usage{
tbk_read_chunk(in_file, hdr, idx_addr, all_units = FALSE)
}
\arguments{
\item{in_file}{the file}

\item{hdr}{the header}

\item{idx_addr}{the index address}

\item{all_units}{read all units? (FALSE)}
}
\description{
read a chunk of a tabixmate file
}
