% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_read_unit.R
\name{tbk_read_unit}
\alias{tbk_read_unit}
\title{read tabixmate units}
\usage{
tbk_read_unit(in_file, idx_addr, hdr, all_units = FALSE)
}
\arguments{
\item{in_file}{filename}

\item{idx_addr}{index addresses}

\item{hdr}{tabixmate header}

\item{all_units}{all units? (FALSE)}
}
\description{
read tabixmate units
}
