% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbk_write_unit.R
\name{tbk_write_unit}
\alias{tbk_write_unit}
\title{write tabixmate chunks}
\usage{
tbk_write_unit(out_file, d1, d2, dtype)
}
\arguments{
\item{out_file}{where to write it}

\item{d1}{data 1}

\item{d2}{data 2}

\item{dtype}{data type (see \code{dtype})}
}
\description{
write tabixmate chunks
}
